% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_department.R
\name{hist_department}
\alias{hist_department}
\title{Display Histogram of the Ages of Desired Department}
\usage{
hist_department(x, y)
}
\arguments{
\item{x}{A character string of desired department. See for_department for a complete
list of departments to use as parameters.}

\item{y}{A numerical value of the number of breaks desired in the histogram. Recommended
amount of breaks: 5-20. Default is 15.}
}
\value{
The average age of Williams faculty for the specified department.
}
\description{
The "hist_department" function produces a histogram of the ages of all Williams
faculty in a particular department and includes an abline of the average age of all
the Williams faculty in the department. Requires the dplyr package for use.
}

