% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_faculty.R
\name{mean_faculty}
\alias{mean_faculty}
\title{Estimate average age of Williams Faculty}
\usage{
mean_faculty(x)
}
\arguments{
\item{x}{The desired year of Williams Faculty. Acceptable arguments are
'clean_df2011', 'clean_df2012', 'clean_df2013', 'clean_df' for data from years
2011-2014. 'clean_df' is default}
}
\value{
The average age of Williams faculty. Default is 2014.
}
\description{
The mean_faculty function estimates the average age of Williams faculty by
year. Available years are 2011-2014.
}

