% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_department.R
\name{summary_department}
\alias{summary_department}
\title{Gives summary of the ages of Williams faculty by department and year}
\usage{
summary_department(x, y)
}
\arguments{
\item{x}{Character string of desired department. See "for_department" for a complete
list of departments to use as parameters. If nothing is entered, "summary_department" will
summarize all departments.}

\item{y}{Data frame to be used. Data frames vary by year. Acceptable parameters include
"clean_df2011", "clean_df2012", "clean_df2013", "clean_df" (years 2011-2014). Note that
"y = clean_df2011" (or 2012/2013) must be entered as argument if user desires summarizing
all departments but does not want the default year 2014.}
}
\value{
A summary of the ages of Williams faculty for the specified department and year.
This function uses the base summary() and sd() functions which will return the mean,
first quartile, median, mean, third quartile, maximum value, and standard deviation.
}
\description{
The "summary_department" function givess a summary of the ages of Williams faculty sorted
by department and year. If no year is specified, 2014 is defaulted.
}

