% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_department.R
\name{mean_department}
\alias{mean_department}
\title{Estimate average age by department}
\usage{
mean_department(x)
}
\arguments{
\item{x}{Character string of desired department. See "for_department" for a complete
list of departments to use as parameters.}
}
\value{
The average age of Williams faculty for the specified department.
}
\description{
The "mean_department" function estimates the average age of Williams faculty sorted
by department.
}

