% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_faculty.R
\name{mean_faculty}
\alias{mean_faculty}
\title{Estimate average age of Williams Faculty}
\usage{
mean_faculty()
}
\arguments{
\item{x}{The desired year of Williams Faculty, (2011, 2012, 2013, or 2014)}
}
\value{
The average age of Williams faculty. Default is 2014.
}
\description{
The mean_faculty function estimates the average age of Williams faculty by
year. Available years are 2011-2014.
}

